extrace gracefully_get_plex
