+++ rarbg_local/new.py
@@ -1,7 +1,6 @@
 import logging
 import os
 import traceback
-from asyncio import wait_for
 from collections.abc import AsyncGenerator, Callable, Coroutine
 from contextvars import ContextVar
 from functools import wraps
@@ -20,7 +19,6 @@ from fastapi.openapi.utils import get_openapi
 from fastapi.requests import Request
 from fastapi.responses import RedirectResponse, Response, StreamingResponse
 from fastapi_utils.openapi import simplify_operation_ids
-from plexapi.server import PlexServer
 from pydantic import BaseModel
 from sqlalchemy import Row, func
 from sqlalchemy.ext.asyncio import AsyncSession, async_object_session
@@ -66,7 +64,7 @@ from .models import (
     TvSeasonResponse,
 )
 from .monitor import monitor_ns
-from .plex import get_imdb_in_plex, get_plex
+from .plex import get_imdb_in_plex, gracefully_get_plex
 from .providers import (
     get_providers,
     search_for_tv,
@@ -363,34 +361,6 @@ async def discover() -> Discover:
 @api.get('/tmdb/configuration')
 async def tmdb_configuration() -> Configuration:
     return await get_configuration()
-
-
-class Appender(Handler):
-    def emit(self, record: logging.LogRecord) -> None:
-        appender = local_appender.get(None)
-        if appender is not None:
-            appender.append(record)
-
-
-logging.getLogger().addHandler(Appender())
-
-
-async def gracefully_get_plex(request: Request, settings: Settings) -> PlexServer:
-    records: list[logging.LogRecord] = []
-    try:
-        token = local_appender.set(records)
-        return await wait_for(get_plex(request, settings), timeout=25)
-    except Exception as exc:
-        local_appender.reset(token)
-        logger.exception('Error getting plex server', exc_info=exc)
-        raise HTTPException(
-            500,
-            {
-                'error': 'Error getting plex server',
-                'details': str(exc),
-                'records': [record.getMessage() for record in records],
-            },
-        ) from exc
 
 
 @api.get('/plex/{thing_type}/{tmdb_id}')
