+++ rarbg_local/websocket.py
@@ -12,11 +12,12 @@ from .db import (
     User,
 )
 from .models import ITorrent, PlexMedia, PlexResponse
-from .plex import get_imdb_in_plex, get_plex
+from .plex import get_imdb_in_plex, gracefully_get_plex
 from .providers import (
     search_for_movie,
     search_for_tv,
 )
+from .settings import get_settings
 from .singleton import get
 from .tmdb import (
     get_movie_imdb_id,
@@ -174,8 +175,8 @@ async def websocket_stream(websocket: WebSocket) -> None:
     elif request.request_type == 'ping':
         message = 'Pong'
     elif request.request_type == 'plex':
-        plex = await get(websocket.app, get_plex, make_request(websocket, request))
-        # plex = await gracefully_get_plex(request, settings)
+        settings = await get(websocket.app, get_settings)
+        plex = await gracefully_get_plex(request, settings)
         dat = await get_imdb_in_plex('movie', request.tmdb_id, plex)
 
         if not dat:
